% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impcheckdata.R
\name{impcheckdata}
\alias{impcheckdata}
\title{Import, control and format your dataset}
\usage{
impcheckdata()
}
\description{
This function is used to import your dataset,
to verify that it contains all the necessary information and
to format it into a suitable data.frame that will be analyzed by the other functions of the package.
}
\details{
Cette fonction sert à importer votre jeu de données,
à vérifier qu'il contient bien toutes les informations nécessaires et
à le formatter en une data.frame adaptée qui sera analysée par les autres fonctions du package.
}
\keyword{dataset}
\keyword{formatting}
\keyword{import}
\keyword{validation}
