% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilitymap.R
\name{mobilitymap}
\alias{mobilitymap}
\title{Map of the mobility network}
\usage{
mobilitymap(x,y)
}
\arguments{
\item{x}{the data.frame containing the degree (mobility parameter) and the GPS coordiantes (column names: LONGITUDE_X and LATITUDE_Y) \cr
le data.frame contenant le degré (paramètre de mobilité) et les coodonnées GPS (nom des colonnes : LONGITUDE_X et LATITUDE_Y)}

\item{y}{(par defaut, 0.8) the size of the localities. It can be a parameters contained in x.}
}
\description{
This function produces a map illustrating the mobility network.
}
\details{
Cette fonction sert à produire une carte de mobilité.
}
\keyword{ggplot}
\keyword{plot}
\keyword{réseau}
\keyword{sna}
