% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobinternat.R
\name{mobinternat}
\alias{mobinternat}
\title{Transboundary movements network}
\usage{
mobinternat(x, mon_pays)
}
\arguments{
\item{x}{the data.frame containing the rawdata and generated by the impcheckdata function. \cr
Required columns: "ORIGINE_NOM", "ORIGINE_PAYS", "ORIGINE_LONGITUDE_X", "ORIGINE_LATITUDE_Y",
"DESTINATION_NOM", "DESTINATION_PAYS", "DESTINATION_LONGITUDE_X", "DESTINATION_LATITUDE_Y" and "EFFFECTIF"}

\item{mon_pays}{the reference country between quotation marks, with the same writting than in x.}
}
\description{
This function analyzes the dataset generated by the impcheckdata function and
produces an international mobility network with respect to a given country.
This network contains movements between the reference country and other countries.
}
\details{
Cette fonction analyse le jeu de données généré par la fonction impcheckdata
et produit un réseau de mobilité internationale par rapport à un pays donné.
Ce réseau contient les mouvements entre le pays de référence et d'autres pays.
}
\examples{
mobinternat(rawdata, "TUNISIE")

}
\keyword{international}
\keyword{movements}
\keyword{network}
