% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobnation.R
\name{mobnation}
\alias{mobnation}
\title{National movements network}
\usage{
mobnation(x, mon_pays)
}
\arguments{
\item{x}{the data.frame containing the rawdata and generated by the impcheckdata function. \cr
Required columns: "ORIGINE_NOM", "ORIGINE_PAYS", "ORIGINE_LONGITUDE_X", "ORIGINE_LATITUDE_Y",
"DESTINATION_NOM", "DESTINATION_PAYS", "DESTINATION_LONGITUDE_X", "DESTINATION_LATITUDE_Y" and "EFFFECTIF"}

\item{mon_pays}{the reference country between quotation marks, with the same writting than in x.}
}
\description{
Function to analyze the dataset generated by the impcheckdata function and
produces a national mobility network with respect to a given country.
This network contains movements inside a given country.
}
\details{
Cette fonction analyse le jeu de données généré par la fonction impcheckdata
et produit un réseau de mobilité nationale par rapport à un pays donné.
Ce réseau contient les mouvements ayant lieu à l'intérieur du pays de référence.
}
\examples{
mobnation(rawdata, "TUNISIE")

}
\keyword{movements}
\keyword{national}
\keyword{network}
