% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramSNA.R
\name{paramSNA}
\alias{paramSNA}
\title{SNA parameters}
\usage{
paramSNA(x, y, mon_pays)
}
\arguments{
\item{x}{the network generated by one of the mobinternat and mobnation functions.}

\item{y}{the data.frame containing the rawdata and generated by the impcheckdata function. \cr
Required columns: "ORIGINE_NOM", "ORIGINE_PAYS", "ORIGINE_LONGITUDE_X", "ORIGINE_LATITUDE_Y",
"DESTINATION_NOM", "DESTINATION_PAYS", "DESTINATION_LONGITUDE_X", "DESTINATION_LATITUDE_Y" and "EFFFECTIF"}

\item{mon_pays}{the reference country between quotation marks, with the same writting than in y.}
}
\description{
This function returns a data.frame with four mobility parameters computed for each locality:
the degree centralities (degree, in-degree, and out-degree) and the betweenness centralities.
}
\details{
Cette fonction 4 paramètres de mobilité pour chaque localité :
le degré, le degré entrant et le degré sortant et la centralité d'intermédiarité.
}
\keyword{network}
\keyword{parameters}
\keyword{sna}
