% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata.R
\docType{data}
\name{rawdata}
\alias{rawdata}
\title{Fake dataset for animal movements between Tunisia and other countries}
\format{An object of class dataframe.}
\usage{
data(rawdata)
}
\description{
Random data generated as an example of inputs required for the analysis.
}
\examples{
head(rawdata)
}
\keyword{datasets}
